[FRONT] fix: build + extract fiches CTA
