+++ application/app/components/fiches/ficheCommune.tsx
@@ -3,7 +3,6 @@ import { useState } from 'react';
 import { Commune } from '@/app/models/communes';
 
 import NbEtablissements from './NbEtablissements';
-import AccordionCard from './shared/AccordionCard';
 import ActionButtons from './shared/ActionButtons';
 import CollectiviteHeaderCard from './shared/CollectiviteHeaderCard';
 import PotentielSolaireCard from './shared/PotentielSolaireCard';
@@ -83,8 +82,6 @@ export default function FicheCommune({ commune }: FicheCommuneProps) {
 					<TopCard topEtablissements={commune.top_etablissements_primaires} />
 				</>
 			)}
-			<br />
-			<AccordionCard />
 		</div>
 	);
 }
