+++ application/app/components/fiches/ficheEtablissement/ficheEtablissement.tsx
@@ -1,7 +1,5 @@
 import { Etablissement } from '@/app/models/etablissements';
 
-import { ELU_BODY, PARTICULIER_BODY, PARTICULIER_END } from '../../content/accordion-actions';
-import AccordionCard from '../shared/AccordionCard';
 import ActionButtons from '../shared/ActionButtons';
 import PotentielSolaireCard from '../shared/PotentielSolaireCard';
 import EtablissementCard from './EtablissementCard';
@@ -9,22 +7,6 @@ import EtablissementCard from './EtablissementCard';
 import InstallationCard from './InstallationCard';
 import InterpretationMessage from './IntepretationMessage';
 import ProtectionCard from './ProtectionCard';
-
-const actionsShort = [
-	{
-		title: 'Je suis un élu et je veux agir',
-		content: <>{ELU_BODY}</>,
-	},
-	{
-		title: 'Je suis un particulier et je veux agir',
-		content: (
-			<>
-				{PARTICULIER_BODY}
-				{PARTICULIER_END}
-			</>
-		),
-	},
-];
 
 interface FicheEtablissementProps {
 	etablissement: Etablissement;
@@ -50,8 +32,6 @@ export default function FicheEtablissement({ etablissement }: FicheEtablissement
 				<InstallationCard surfaceExploitableMax={etablissement.surface_exploitable_max} />
 				{/* <GraphiqueCard /> */}
 			</div>
-			<hr className='my-2' />
-			<AccordionCard actions={actionsShort} />
 		</div>
 	);
 }
